refactor: Modify executeCommand method parameters to accept a List of targetServerIds
This commit refactors the executeCommand method in the CommandExecutor class to accept a List of targetServerIds instead of a single targetServerId. This change allows for executing a command on multiple servers simultaneously. The method now iterates over the list of targetServerIds and executes the command accordingly. Additionally, the scheduleCommandExecution method in the CommandRegistrar class has been updated to handle the List of targetServerIds.