feat: Add support for parsing command arguments in StringParser
Added support for parsing command arguments in the `StringParser` class by adding a new parameter `String[] args` to the `parsePlaceholders`, `parseConsoleCommands`, and `parseBlockCommands` methods. Also, introduced a new private method `appendArguments` to concatenate the arguments into a single string. This enhancement improves the flexibility and functionality of the command parsing process.